python_web/provider/pyproject.toml
[project]
name = "provider"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.12.*"
dependencies = [
    "backend-common",
    "aiokafka>=0.10.0",
    "asyncpg>=0.30.0",
    "aiofiles>=25.1.0",
    "fastapi>=0.135.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "alembic>=1.15.0",
    "pydantic[email]>=2.12.5",
    "pydantic-extra-types>=2.11.0",
    "phonenumbers>=9.0.26",
    "ruff>=0.15.9",
]

[tool.uv.sources]
backend-common = { path = "../../shared", editable = true }

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
